test(cloud-ai-sdk): delete duplicate and tautological unit tests; fold load-cancellation invariant into contract suite
