+++ packages/cloud-ai-sdk/src/core/CloudChatCore.test.ts
@@ -85,74 +85,4 @@ describe("CloudChatCore", () => {
     expect(onToolCall).toHaveBeenCalledWith({ toolCall: {} });
     expect(result).toBe(completion);
   });
-
-  it("generates a title once for a newly created thread after assistant output", async () => {
-    const generateTitle = vi.fn().mockResolvedValue("Generated title");
-    const core = createCore({ generateTitle });
-
-    const meta = { threadId: "thread-1", loading: null, loaded: false };
-    const registry = {
-      getMeta: vi.fn().mockReturnValue(meta),
-      getOrCreateMeta: vi.fn().mockReturnValue(meta),
-      get: vi.fn().mockReturnValue({
-        messages: [{ id: "m-1", role: "assistant" }],
-      }),
-      getOrCreate: vi.fn(),
-    } as never;
-
-    core.titlePolicy.markNewThread("thread-1");
-    await core.persistChatMessages("chat-1", registry);
-    await core.persistChatMessages("chat-1", registry);
-
-    expect(persistMock).toHaveBeenCalled();
-    expect(generateTitle).toHaveBeenCalledTimes(1);
-    expect(generateTitle).toHaveBeenCalledWith("thread-1");
-  });
-
-  it("routes load errors to onSyncError and clears loading state", async () => {
-    const onSyncError = vi.fn();
-    const failure = new Error("load failed");
-    loadMessagesMock.mockRejectedValue(failure);
-
-    const core = createCore({ onSyncError });
-
-    const meta = {
-      threadId: "thread-1",
-      loading: Promise.resolve(),
-      loaded: false,
-    };
-    const registry = {
-      getOrCreateMeta: vi.fn().mockReturnValue(meta),
-      getOrCreate: vi.fn().mockReturnValue({ messages: [] }),
-    } as never;
-
-    await core.loadThreadMessages("thread-1", "chat-1", registry, {
-      cancelled: false,
-    });
-
-    expect(onSyncError).toHaveBeenCalledWith(failure);
-    expect(meta.loading).toBeNull();
-  });
-
-  it("does not clear a replacement load after a cancelled load fails", async () => {
-    loadMessagesMock.mockRejectedValue(new Error("cancelled load"));
-
-    const core = createCore();
-    const replacementLoad = Promise.resolve();
-    const meta = {
-      threadId: "thread-1",
-      loading: replacementLoad,
-      loaded: false,
-    };
-    const registry = {
-      getOrCreateMeta: vi.fn().mockReturnValue(meta),
-      getOrCreate: vi.fn().mockReturnValue({ messages: [] }),
-    } as never;
-
-    await core.loadThreadMessages("thread-1", "chat-1", registry, {
-      cancelled: true,
-    });
-
-    expect(meta.loading).toBe(replacementLoad);
-  });
 });
